Move presenters to lib
