+++ lib/compendium/presenters.rb
@@ -1,0 +1,12 @@
+module Compendium
+  module Presenters
+    autoload :Base,     'compendium/presenters/base'
+    autoload :Chart,    'compendium/presenters/chart'
+    autoload :CSV,      'compendium/presenters/csv'
+    autoload :Metric,   'compendium/presenters/metric'
+    autoload :Option,   'compendium/presenters/option'
+    autoload :Query,    'compendium/presenters/query'
+    autoload :Table,    'compendium/presenters/table'
+    autoload :Settings, 'compendium/presenters/settings'
+  end
+end
