+++ lib/compendium/presenters/chart.rb
