+++ lib/compendium/presenters/settings.rb
@@ -1,0 +1,8 @@
+module Compendium
+  module Presenters
+    module Settings
+      autoload :Query, 'compendium/presenters/settings/query'
+      autoload :Table, 'compendium/presenters/settings/table'
+    end
+  end
+end
