+++ spec/compendium/presenters/base_spec.rb
