+++ spec/compendium/presenters/settings/query_spec.rb
