+++ spec/compendium/presenters/settings/table_spec.rb
