+++ spec/compendium/presenters/table_spec.rb
